Image optimisation and some JS optimisation
Ensure all images are lazy loaded (added JS library for this), budled and minimised the footer scripts and loaded async, used <picture> elements with source sets to load webp versions if possible, removed WebPFallbackImageUrl image helper to ensure the site can be cached on a CDN if appropriate. Tested on common browsers. Images don't show on IE but content is visible which is the most important thing for the site